fixed error detecting the same app.
it took all day but I finally fixed the error where it keeps finding and
adding all your apps to the newly installed list on every boot.
However because of the new method, the old method for detecting
uninstalled apps is now unreliable and will have to be remade based on
this new method.