various minor fixes, removed unused class, made an async task
Fixed a bunch of typos everywhere.
Removed the subAppDetail class that was unused and won't be used later.
Updated the getDrawable functions to ones that aren't depreciated.
Created an Async task to handle unimportant things that take excess
time, like loading images (Yes I know this was meant for RC2, but I
rushed it because I could).
Updated Readme.md