NEW BUILD. plus Cleanup and fixes.
Removed the remainder of the text glow functionality, and replaced it
with a text shadow and typeface plus an icon scale, which removes a LOT
of TextViews that would otherwise be drawn.
Removed the icon background, something that works as a background for
the view of the entire item would be better.
Removed some redundant code on creating menu entries.
Made the area for menu items larger to add room for more text.
Cleaned out some unused includes.
Updated READEME.md
Updated save file structure to add all currently expected fields for
data, to help reduce the chance of future save file corruption.
Last but not least, it seems stable enough for a new build! So here's a
stable revision!