Bugfixes, UI improvements, and Android 6
This update is a bit late because I had to do a complete system
reinstall.

Rebuilt some UI code, should be a bit faster to load up, if nothing
else, there's less code to make it load.
Modified where the apps list appears to line it up better.
Fixed a occasional crash due to a formatting error in the main layout.
Fixed the options menu not closing when you remove an item from a list.
Cleaned up some code, just to make it a little easier to read.
Updated target API to Android 6. still works all the way back to 4.0.3.
But all the newer devices seem to be moving to 6 without hesitation, so
it will need better support.
Updated README.md

This update marks the final build of RC1. All the bugs I could find with
thorough testing have been patched and all the core features are in
place.
From here, development of RC2 will begin.