bugfix and Implemented color select
Fixed a small glitch that sometimes happen when the app resumes after
uninstalling an app.
Added a new layout for the color selector.
Implemented a color picker, that will be later used to select the color
for the font (and later icons and menus, once the images are done).
Updated the save file to support that colors are now integers, doesn't
seem to break old save files since their previous value was 0 anyway.
Color picker currently will let you change the variable for the font
color, however the variable is not yet used to actually modify the font.