Code cleaning
Created a second script to simplify code.
A number of functions have been taken from the main activity's script
and offloaded to a different script, this is in an attempt to reorganize
the code. With the rapid growth rate of the code, hitting nearly 1000
lines in a single script, it will be very difficult to manage without
doing this to some degree or another.
I will also start trying to partition off different areas of code with
larger comments to make it easier to skim to what you want.
Next update should have the remaining things moved over into their
proper places.