Cleanup, bugfixes, and a new build!
Yes this one is tested.
The hli variable was moved to a list of Views, instead of a list of
AppDetails, makes it easier to manage, and removes a function that
shouldn't be needed. This also fixes a small bug created with the
refactoring, unexpected side bonus; it seems to save some RAM.
Fixed a glitch with categories not being properly selected, that was due
to the refactoring.
Fixed a glitch where the "go back" button on a menu would open the wrong
menu, due to refactoring.
Fixed glitch where the "exit menu" button didn't close the menu due to
refactoring.
Improved the detection for if you are moving out of bounds from the
categories.
fixed a pre-existing potential crash, or at least selection graphic
error when changing categories.
Changed the font, now the shadow will always show, it is always
anti-alaiased, and bold is now set through Paint. this new method also
allows these variables to be changed without instancing the TextView to
hopefully save some RAM.
New alpha build for RC2 with the new test feature (color changer), and
the new refactoring/optimizations.