Bugfixed, Organization method selection, and more
The new changes made to settings file in this update do not seem to
cause instability.
Added new icons (from Google Material Design) to show which item is
selected.
Added a new variable to the save file to track how you organize each
menu.
Added Alphabetical organize option.
Updated README.md
Removed a log call that is no longer needed.
Added Reverse Alphabetical organize option.
Added Do not organize option.
Organization can either be done once, or every time the menu is loaded.
Removed an unused variable in appDetail.
Added a persistent option in AppDetail, this is used to bypass the item
removal check, this is necessary for custom items that link back to
LaunchIntents not listed in the system ResolveInfo (a web bookmark or a
PPSSPP game as examples), which support will be added later on.