Minor tweaks and a new build.
Reorganized some code.
Removed a couple useless variable creations.
Updated README.md, I need to prioritize management of save data, even
the current system has some bugs, and the only way I can think to
overcome it is using XML.
Made the font slightly bigger and changed the shadow from a glow to a
soft shadow, spent the entire day experimenting with text to improve
readability and that was the best I found.
New build with the changes.