working new save format
Added function to request permission for write storage if it is needed,
this will only appear in android 6 and later due to the way 6+ handles
security.
Removed some now unnecessary logging.
Made some modifications to the string generator, but now it will
generate a save file in XML format.
Tested with 107 apps installed it only takes about 52 thousand
characters out of a limit of 2 billion (java limitation on string
length), so save size won't be an issue even for pack rat users.
It also appears we can use this to write the new save file format, while
still using the old format at the same time, so revision 2 may allow for
transitioning save files between RC2 and RC2.5+.