Shadow changes, save data progress,
!!TESTING CODE!!
Most of the new things in this is heavily unfinished and very
unoptimized.
Improved the organization of the save file that get created.
Added some extra data that gets put into the new save format, to support
an upcoming feature (custom categories).
Fixed an issue with the & symbol creating read errors in the save files.
Fixed an issue with the XML not having a root element.
Added a portion of the necessary XML parse, I'm using methods intended
for parsing HTML because it's a lot more efficient, and simpler, than
any of the java/android XML managers.