!!UNTESTED!! More progress to RC2 Final.
!! The changes in this update have not been tested !!
Changed the hList to a list of AppDetails, this will allow us to change
it later on, adding custom categories and such, this basically meant
rebuilding everything related to it.
Added an image loader class, since custom categories will require an
obscene list of images that may or may not be loaded, I made a custom
class to handle that later, right now it's more of a placeholder.
Added some more documentation.
Updated README.md