performance improvements
!! Changes to this update have not been tested !!
The oldApps list now only has value when it's being changed, the
savedata.oldApps variable still maintains the value, but overall this
will save some RAM since we no longer need to constantly maintain two of
the same variable.
The app's data is now set to save in the data directory for the app, we
will have an option to save to sdcard0/sdcard1/usb1/etc as a backup
setting in options menu later, we can also give an option for automatic
backups to that location.
Updated README.md
The organizeMode is now going to be saved and loaded as part of the
hList, since they work hand in hand.
Added loading for hiddenApps in the new format.
Fixed loading for oldApps in the new format.