Bugfixes and a new build.
I am so glad to be done with RC2. You have no idea.
Uploaded the new build finally! Yes it has been tested, and seems to do
its job well enough.

Promoted creating the hMenu to a function of its own, to let me call it
and create the menu before I save, avoids a lot of errors when saving
until the new hMenu loader can be implemented in 2.5.
Reverted the scrolling bugfix, apparently it was right before.
Changed the method for the new save file to a FileOutputStream, I can
only write to internal memory this way, I left the old method there as a
note because I will need it for the backup function.
Fixed a glitch where it fails to scroll to the hMenu on start and causes
a crash.
Rebuilt the method for creating the new hMenu.
Fixed a glitch where clicking a category did nothing.
Partially fixed a glitch where long clicking a category opens options
(still happens randomly but doesn't seem constant now).
Reorganized some whitespacing.
Fixed a glitch in the save files when it's trying to detect the New Apps
category when empty (this fix will be useless and removed in 2.5).
Made the list of apps wider to support longer names.
Added some missing positioning data for the button on category items.