Beginning of 2.5
The code is tested and runs, but I still don't trust everything is going
as smoothly as it seems.
Disabled saving so I don't accidentally break my own save file in the
process of getting the ability to read the save file working.
Cleaned out some imports.
Fixed a crash when no save file was found.
Program now mostly reads the save file.
Added some debugging on keypress in attempt to later fix some
incompatibilities with the hyperkin PS1/2 USB adapter.
organizeMode is now a List<int[3]> instead of a int[7][3].
Fixed an organizational typo in the save file (it had no effect besides
being ugly to read).
Updated README.md to mark we are now on version 2.5.