Bugfixes and API update. New build.
Updated the API to 23.0.2.
Fixed an error with some devices when writing the save file to the
internal memory.
Improved the saving method in the process, it should be more efficient
now.
Fixed a crash when navigating the options menu via keyboard/controller.
Fixed a glitch with being able to scroll out of the categories list with
keyboard/controller.
Fixed a scrolling issue with keyboard/controller.
New build that has the fixes.