Bugfixes, performance tweaks, and new builds!
Fixed an occasional glitch where uninstalled apps don't always get
removed, the new method may degrade performance to a degree in exchange
for accuracy.
The check for newly installed apps now runs through the same loop as the
check for removed apps, to improve performance.
The check for in-built apps now runs through a small loop in the check
for removed apps, to very slightly improve performance and accuracy.
All the changes allowed me to obsolete the oldApps variable to save RAM
and storage space.
New build with the improvements.