Background dim, theme changes, and performance improvements
Added a new feature to optionally dim the list views.
Added the switch to enable/disable the list view dim to the options
menu.
Initialization now has it's own class that runs on an async thread, this
seems to improve performance for the startup significantly.
Added a small animation to be used as a delay in processing.
To improve the speed that the app opens, an animation was also used in
the process of setting up the initialization to act as a small delay,
allowing the application to load up before the initialization has to
start.
Changed some various sizes in the layouts to improve performance, make
icons easier to see, and text easier to read.
Fixed a glitch where icons did not display properly when re-sized.
New build that contains the new features.