Bugfixes, inbuilt toolbar, category tags, and accessory detection
Added menu tags, this will help define how google apps are auto
organized into lists later, and helps define what menus to jump to on
accessory change.
Implemented a function to detect when you plugin a wired headset, and
change the menu to the menu with the Music tag and display a toast that
you plugged in a headset.
Cleaned the initialization script, initial variable values are now fully
managed by the save data loader.
Reorganized the category and list items a little.
Refactored the AppDetail to start with a lowercase letter.
Fixed some missing break calls in switch classes during save file load.
Added the start of the main toolbar at the top of the app, which will
replace the notification bar.
The function for detecting if a headset is plugged in also works for
managing the clock in the toolbar.
Added a search bar and button to the toolbar, currently they do nothing
significant.
Shortened the delay animation to improve performance and help manage the
clock start.
Modified the stray image view in the main layout, later it will function
as the custom image background for apps that have one.