searchbar functionality, drawables removed, and other changes
The searchbar is now able to display results of installed apps.
The listdim xml has been moved to a xml list of colors used.
Images added for the searchbar.
Drawables have been moved to Bitmaps, in an attempt to make the Garbage
Collector actually do something every once in a while.
Fixed error when it tries and fails to load the Google icon for
extras/google drive.
Finally tossed the now obsolete svgLoad function that never lived up to
its intended purpose.
Re-worked the layouts again.
Cleaned out a bunch of imports.