Bugfixes and misc changes.
Removed an unnecessary clear of a list on app start (potentially
improving performance in some rare cases).
Updated the function for adding built-in apps to the lists, now it will
detect what category they are supposed to be in when adding, via the
tags.
Fixed the double save on start glitch, I think it has been fixed and I'm
just finding out late.
Uploaded the missing search category layout, for some reason it didn't
get uploaded in the last update.
Fixed a glitch when an app would sometimes be opened after releasing the
DPAD.
Simplified the listMove function to improve performance.
Refactored some names to be easier to read.
Improved the method of processing initialization, it should use proper
threading methods now to improve performance and to trim down the
initialization script a little.
Fixed a glitch for pre-existing users where the tags didn't appear
properly (this will have to be removed later since the glitch wont
effect new users and the fix will hinder later revisions).
Added a class for searching the web.
The searchbar can now search the web (google's ajax interface), although
for the moment the results are only displayed in the debug log.
Updated the README.md