Bugfixes, new build!
Fixed tags not saving.
Fixed tags being initialized improperly when not found in the save file.
Fixed a glitch with the IntentReciever not being unregistered when you
leave the app.
Slightly improved performance when detecting if an app is installed.
Fixed a rare glitch where some removed apps aren't removed until the
second resume.
Fixed the double save glitch for sure this time.
Reorganized initial loading function to be easier to read, and
potentially improve performance slightly.
Added a new 2.5 build with the bugfixes and the new tag functionality.