Cleanup and optimizations
Removed the unused raw images (the art is getting replaced anyway and we
won't be using svg)
Changed the type of AsyncTask used tor initialization to try and help
improve performance.
Moved and changed when loadApps is called to try and improve
performance.
Changed the OptionsMenuChange functions to no longer require a view
variable being thrown around to improve loading speeds and save RAM.
Fixed a small loop error in the save file loader.
Added compensation in the save file loader in case of read error.
Finally fixed reverse alphabetical organization.
Attempted optimizations to the method for creating the save file, more
to come if this works.