April Fools
Honestly, I really don't trust the current stage, everything looks right
but this thing is doing anything but gaining my trust. Which is why I'm
Posting the builds as Alpha stage again.
So here's a new alpha build.
The prank is that it's probably one of the most stable builds of 2.5
thus far.