Bugfixes, new build
Finally fixed the memory leak, I'm assuming it was due to WHEN I was
saving, not how.
Fixed the cut-off of icons when an app is selected.
Fixed some text that is larger than than the textbox.
App memory is now cleaned when it goes into the background.
onPause is now executed in proper order.
The initialization script has been rebuilt a but to be more efficient,
and is no longer async to help avoid crashing on weaker devices.
Removed some (now) useless functions and documentation.
Save function now only needs one variable to call it.
new build with the changs