Refactoring, bugfixes, and miner optimizations, new build
Refactored the variable classes to their own folder(organizational
purposes only).
Improved the use of bools in the onClick function for list entries.
The Image View in the menu entries is no longer instanced during
creation, to try and improve performance.
Fixed a glitch with not being able to use internal icons on the
categories.
Fixed a glitch/crash with controllers where you were able to scroll out
of the list.
Rebuilt the function for the list scroll with controllers to be easier
to read.
New build with the changes.